Add files via upload
This is an example of a data analysis engine developed for web traffic analysis on a variety of scales.  This script represents all of the stored procedures and functions needed for the base engine.  Specific business requirements may be added depending on the addition of new source systems or new business rules.